Long ProgressRequirement

+++ src/main/kotlin/net/trueog/questsOG/Requirement.kt
@@ -2,6 +2,6 @@ package net.trueog.questsOG
 
 open class Requirement(val name: String)
 
-class ProgressRequirement(name: String, val current: Int, val target: Int) : Requirement(name)
+class ProgressRequirement(name: String, val current: Long, val target: Long) : Requirement(name)
 
 class BooleanRequirement(name: String, val met: Boolean) : Requirement(name)
